+++ src/components/DpFlyout/DpFlyout.vue
@@ -6,15 +6,17 @@
       'o-flyout--right': align === 'right',
       'o-flyout--padded': padded,
       'is-expanded': isExpanded,
-      'o-flyout--menu': hasMenu
+      'o-flyout--menu': hasMenu,
+      'bg-surface-medium rounded-md': variant === 'dark'
     }"
     data-cy="flyoutTrigger">
     <button
       :disabled="disabled"
       type="button"
       aria-haspopup="true"
-      class="o-flyout__trigger btn--blank o-link--default u-ph-0_25 line-height--2 whitespace-nowrap"
-      :data-cy="dataCy"
+      :aria-label="ariaLabel !== '' ? ariaLabel : null"
+      class="o-flyout__trigger btn--blank o-link--default px-1 line-height--2 whitespace-nowrap"
+      :data-cy="dataCy !== '' ? dataCy : null"
       @click="toggle">
       <slot
         name="trigger"
@@ -48,6 +50,12 @@ export default {
       validator: (prop) => ['left', 'right'].includes(prop)
     },
 
+    ariaLabel: {
+      type: String,
+      required: false,
+      default: ''
+    },
+
     dataCy: {
       type: String,
       required: false,
@@ -70,7 +78,14 @@ export default {
       required: false,
       type: Boolean,
       default: true
-    }
+    },
+
+    variant: {
+      required: false,
+      type: String,
+      default: 'light',
+      validator: (prop) => ['light', 'dark'].includes(prop)
+    },
   },
 
   data () {